+++ pkg/validator/json.go
@@ -11,24 +11,22 @@ type JSONValidator struct{}
 
 var _ Validator = JSONValidator{}
 
-func getCustomErr(input []byte, err error) error {
-	var jsonError *json.SyntaxError
-	if !errors.As(err, &jsonError) {
-		return err
-	}
-
-	offset := int(jsonError.Offset)
-	line := 1 + strings.Count(string(input)[:offset], "\n")
-	column := 1 + offset - (strings.LastIndex(string(input)[:offset], "\n") + len("\n"))
-	return fmt.Errorf("error at line %v column %v: %w", line, column, jsonError)
-}
-
 func (JSONValidator) ValidateSyntax(b []byte) (bool, error) {
 	var output any
 	err := json.Unmarshal(b, &output)
 	if err != nil {
-		customError := getCustomErr(b, err)
-		return false, customError
+		var synErr *json.SyntaxError
+		if errors.As(err, &synErr) {
+			offset := int(synErr.Offset)
+			line := 1 + strings.Count(string(b)[:offset], "\n")
+			column := 1 + offset - (strings.LastIndex(string(b)[:offset], "\n") + len("\n"))
+			return false, &ValidationError{
+				Err:    fmt.Errorf("error at line %v column %v: %w", line, column, synErr),
+				Line:   line,
+				Column: column,
+			}
+		}
+		return false, err
 	}
 	return true, nil
 }