+++ pkg/validator/plist.go
@@ -2,9 +2,13 @@ package validator
 
 import (
 	"bytes"
+	"regexp"
+	"strconv"
 
 	"howett.net/plist"
 )
+
+var plistLineColRe = regexp.MustCompile(`at line (\d+) character (\d+)`)
 
 // PlistValidator is used to validate a byte slice that is intended to represent a
 // Apple Property List file (plist).
@@ -18,6 +22,11 @@ func (PlistValidator) ValidateSyntax(b []byte) (bool, error) {
 	plistDecoder := plist.NewDecoder(bytes.NewReader(b))
 	err := plistDecoder.Decode(&output)
 	if err != nil {
+		if m := plistLineColRe.FindStringSubmatch(err.Error()); m != nil {
+			line, _ := strconv.Atoi(m[1])
+			col, _ := strconv.Atoi(m[2])
+			return false, &ValidationError{Err: err, Line: line, Column: col}
+		}
 		return false, err
 	}
 	return true, nil