+++ pkg/validator/toon.go
@@ -4,9 +4,13 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"regexp"
+	"strconv"
 
 	"github.com/toon-format/toon-go"
 )
+
+var toonLineRe = regexp.MustCompile(`line (\d+):`)
 
 type ToonValidator struct{}
 
@@ -15,6 +19,11 @@ var _ Validator = ToonValidator{}
 func (ToonValidator) ValidateSyntax(b []byte) (bool, error) {
 	_, err := toon.Decode(b)
 	if err != nil {
+		if m := toonLineRe.FindStringSubmatch(err.Error()); m != nil {
+			if line, convErr := strconv.Atoi(m[1]); convErr == nil {
+				return false, &ValidationError{Err: err, Line: line}
+			}
+		}
 		return false, err
 	}
 	return true, nil